Detect and render GFM task lists.
In the future, we should use npm to keep our markdown-it and friends up to date instead of directly adding them and changing them in our repository. At that time, we would stop using this current implementation and start using one of the available plugins for task lists in markdown-it.